% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{as_module}
\alias{as_module}
\title{Use a Package as if a Module}
\usage{
as_module(package)
}
\arguments{
\item{package}{name of a package; character}
}
\value{
a module that contains a package's exported objects
}
\description{
Use a Package as if a Module
}
\examples{

covr <- as_module("covr")
ls(covr)
\dontrun{
covr$codecov()
}

}
