% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{drop}
\alias{drop}
\title{Detach a Module from the Search Path}
\usage{
drop(name)
}
\arguments{
\item{name}{name of the module to exit from; character}
}
\value{
\code{TRUE} if successful; invisible
}
\description{
If no argument is supplied, detach the most recently attached module.
}
\examples{

use(mod::ule({
   a <- 1
}), as = "my_module")

use(mod::ule({
   b <- 2
}), as = "my_other_module")

search()

# by name
drop("my_module")

# and at the head position
drop()

search()

}
\seealso{
\code{\link{use}}
}
