% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R, R/namespace.R
\name{enter}
\alias{enter}
\title{Enter a module}
\usage{
enter(file, as = paste0(basename(file)), ...)

enter(file, as = paste0(basename(file)), ...)
}
\arguments{
\item{file}{path to an R file}

\item{...}{dot-dot-dot, any additional arguments for 'attach' function}

\item{name}{name of the module to be used in search path}

\item{file}{path to an R file}

\item{name}{name of the module to be used in search path}

\item{...}{dot-dot-dot, any additional arguments for 'attach' function}
}
\description{
Load and attach a module to the search path. This method disregards `provide()` statement.

Load and attach a module to the search path. This method disregards `provide()` statement.
}
