% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R, R/namespace.R
\name{exit}
\alias{exit}
\title{Exit from a module}
\usage{
exit(name)

exit(name)
}
\arguments{
\item{name}{name of the module to exit from}

\item{name}{name of the module to exit from}
}
\description{
Detach a module from the search path.

Detach a module from the search path.
}
