% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{module}
\alias{module}
\title{Declare a module in-line}
\usage{
module(...)
}
\arguments{
\item{...}{expression}
}
\value{
an environment containing objects from the module
}
\description{
Declare a module in-line
}
