% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/declaratives.R
\name{name}
\alias{name}
\title{Name a Module}
\usage{
name(name)
}
\arguments{
\item{name}{the name of the module; character}
}
\value{
the input
}
\description{
Name a Module
}
\examples{

mod::ule({
    name("my")
    # ...
})

}
