% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{private}
\alias{private}
\title{Extract the Private Environment of a Module}
\usage{
private(module)
}
\arguments{
\item{module}{a module}
}
\value{
environment
}
\description{
Extract the Private Environment of a Module
}
\examples{

m <- mod::ule({a <- 1})
pvt <- private(m)

ls(pvt, all.names = TRUE)
}
