% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/declaratives.R
\name{provide}
\alias{provide}
\title{Provide objects from a module}
\usage{
provide(...)
}
\arguments{
\item{...}{dot-dot-dot: name of any object to be accessible by user}
}
\description{
Only legal inside a module context
}
\examples{

\dontrun{
provide(a, c)

a <- 1
b <- 2
c <- 3
d <- 4
}
}
