% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{refer}
\alias{refer}
\title{Exposes the objects from one environment to another}
\usage{
refer(source)
}
\arguments{
\item{source}{the providing environment}
}
\description{
Exposes the objects from one environment to another
}
