% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thing.R
\name{[.thing}
\alias{[.thing}
\title{Invoke the Active Binding in a Thing}
\usage{
\method{[}{thing}(x, ...)
}
\arguments{
\item{x}{a thing}

\item{...}{dot-dot-dot, ignored}
}
\description{
Invoke the Active Binding in a Thing
}
