% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thing.R
\name{thing}
\alias{thing}
\title{Make a Thing}
\usage{
thing(..., dot, parent = parent.frame(), lock = FALSE,
  expose_private = FALSE)
}
\arguments{
\item{...}{module expression}

\item{dot}{function expression used for active binding to `.`}

\item{parent}{the enclosing environment}

\item{lock}{lock the environment; logical}

\item{expose_private}{expose the private environment as `..private..`; logical}
}
\value{
a \code{module} containing an active binding
}
\description{
A "thing" is a special object made based on a module.
Contains an active binding, defined with the `dot` argument.
}
\examples{

my_thing <- mod::thing({
    a <- 1
}, dot = function() a)

my_thing$.

my_thing[]

}
