% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{use}
\alias{use}
\title{Use a module}
\usage{
use(module, as, all_objects = FALSE, ...)
}
\arguments{
\item{module}{path to an R file or a symbol for a module object}

\item{as}{name to be used in the search path}

\item{all_objects}{Boolean; whether to include all objects, disregarding `provide()` declarations}

\item{...}{dot-dot-dot, any additional arguments for 'attach' function}
}
\description{
Load and attach a module to the search path
}
\examples{
\dontrun{
use("~/R/my_module.R")
}
}
