% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{acquire}
\alias{acquire}
\title{Load a module from a file}
\usage{
acquire(file, all_objects = FALSE, parent = .GlobalEnv)
}
\arguments{
\item{file}{path to an R file}

\item{all_objects}{Boolean; whether to include all objects, disregarding `provide()` declarations}

\item{parent}{the enclosing environment}
}
\value{
an environment containing objects from the module
}
\description{
Load a module from a file
}
\examples{
module_path <- system.file("misc", "example_module.R", package = "modular")
example_module <- acquire(module_path)

ls(example_module)

example_module$a
example_module$e(123)

}
