% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{acquire}
\alias{acquire}
\title{Acquire a module from a file}
\usage{
acquire(file, all_objects = FALSE)
}
\arguments{
\item{file}{path to an R file}

\item{all_objects}{Boolean; whether to include all objects, disregarding `provide()` declarations}
}
\value{
an environment containing objects from the module
}
\description{
any hidden object (which name/symbol preceded by `.`) are only accessible through
function defined in the module.
}
