% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{acquire}
\alias{acquire}
\title{Acquire a module from a file}
\usage{
acquire(file, all_objects = FALSE, lock_bindings = TRUE)
}
\arguments{
\item{file}{path to an R file}

\item{all_objects}{Boolean; whether to include all objects, disregarding `provide()` declarations}

\item{lock_bindings}{Boolean; TRUE to prevent values being altered from outside the module.}
}
\value{
an environment containing objects from the module
}
\description{
any hidden object (which name/symbol preceded by `.`) is only accessible through
functions defined in the module.
}
\examples{
my_module_path <- system.file("misc", "my_module.R", package = "modular")
my_module <- acquire(my_module_path)

ls(my_module)

my_module$a
my_module$e(123)

}
