% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{drop}
\alias{drop}
\title{Drop a Module}
\usage{
drop(name)
}
\arguments{
\item{name}{name of the module to exit from; character}
}
\description{
Detach a named module from the search path. If no arguments is supplied, detach the most recently attached module.
}
\examples{

\dontrun{
# by name
drop("my_module")
# at the last position
drop()
}
}
