% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/namespace.R
\name{enter}
\alias{enter}
\title{Enter a module}
\usage{
enter(file, as = paste0(basename(file)), ...)
}
\arguments{
\item{file}{path to an R file}

\item{...}{dot-dot-dot, any additional arguments for 'attach' function}

\item{name}{name to be placed in evaluation search path}
}
\description{
Enter a module
}
