% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{module}
\alias{module}
\alias{thing}
\alias{acquire}
\title{Make a module}
\usage{
module(..., parent = .GlobalEnv, force_public = FALSE, lock = TRUE)

thing(..., force_public = TRUE, lock = TRUE)

acquire(file, parent = .GlobalEnv, force_public = FALSE, lock = TRUE)
}
\arguments{
\item{...}{expression}

\item{parent}{the enclosing environment}

\item{force_public}{force every binding to be public}

\item{lock}{lock the environment}

\item{file}{path to an R file}
}
\value{
an environment containing objects from the module
}
\description{
Make a module
}
\examples{
module_path <- system.file("misc", "example_module.R", package = "modular")
example_module <- acquire(module_path)

ls(example_module)

example_module$a
example_module$e(123)

}
