% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R, R/thing.R
\name{module}
\alias{module}
\alias{bare_module}
\alias{acquire}
\alias{expose}
\alias{thing}
\title{Make a module}
\usage{
module(..., parent = parent.frame(), lock = TRUE,
  expose_private = FALSE)

bare_module(..., parent = baseenv(), lock = TRUE,
  expose_private = FALSE)

acquire(module, parent = .GlobalEnv, lock = TRUE,
  expose_private = FALSE)

expose(module, as, parent = .GlobalEnv, lock = TRUE,
  expose_private = FALSE)

thing(..., dot, parent = parent.frame(), lock = TRUE,
  expose_private = FALSE)
}
\arguments{
\item{...}{module expression}

\item{parent}{the enclosing environment}

\item{lock}{lock the environment}

\item{expose_private}{expose the private environment as `..private..`}

\item{module}{module object, or path to a module file}

\item{as}{name when attached to search path with `use()`}

\item{dot}{function expression used for active binding to `.`}
}
\value{
an environment containing objects from the module
}
\description{
Make a module
}
\examples{
module_path <- system.file("misc", "example_module.R", package = "modular")
example_module <- acquire(module_path)

ls(example_module)

example_module$a
example_module$e(123)

}
