% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{module}
\alias{module}
\alias{thing}
\alias{acquire}
\title{Make a module}
\usage{
module(..., parent = .GlobalEnv, lock = TRUE, expose_private = FALSE)

thing(..., dot, lock = TRUE)

acquire(file, parent = .GlobalEnv, lock = TRUE,
  expose_private = FALSE)
}
\arguments{
\item{...}{expression}

\item{parent}{the enclosing environment}

\item{lock}{lock the environment}

\item{dot}{function expression used for active binding to `.`}
}
\value{
an environment containing objects from the module
}
\description{
Make a module
}
\examples{
module_path <- system.file("misc", "example_module.R", package = "modular")
example_module <- acquire(module_path)

ls(example_module)

example_module$a
example_module$e(123)

}
