% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{module}
\alias{module}
\title{Declare a module in-line}
\usage{
module(..., all_objects = FALSE, lock_bindings = TRUE)
}
\arguments{
\item{...}{expression}

\item{all_objects}{Boolean; whether to include all objects, disregarding `provide()` declarations}

\item{lock_bindings}{Boolean; TRUE to prevent values being altered from outside the module.}
}
\value{
an environment containing objects from the module
}
\description{
Declare a module in-line
}
