% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{module}
\alias{module}
\alias{submodule}
\alias{thing}
\alias{acquire}
\alias{use}
\title{Make a module}
\usage{
module(..., parent = .GlobalEnv, lock = TRUE, expose_private = FALSE)

submodule(..., dot, lock = TRUE, expose_private = FALSE)

thing(..., dot, lock = TRUE, expose_private = FALSE)

acquire(module, parent = .GlobalEnv, lock = TRUE,
  expose_private = FALSE)

use(module, as, parent = .GlobalEnv, lock = TRUE,
  expose_private = FALSE)
}
\arguments{
\item{...}{dot-dot-dot}

\item{parent}{the enclosing environment}

\item{lock}{lock the environment}

\item{expose_private}{expose the private environment as `..private..`}

\item{dot}{function expression used for active binding to `.`}

\item{module}{module object, or file path to a module file}

\item{as}{name when attached to search path with `use()`}

\item{expr}{module expression}
}
\value{
an environment containing objects from the module
}
\description{
Make a module
}
\examples{
module_path <- system.file("misc", "example_module.R", package = "modular")
example_module <- acquire(module_path)

ls(example_module)

example_module$a
example_module$e(123)

}
