% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{print.module}
\alias{print.module}
\title{Print a Module}
\usage{
\method{print}{module}(x, ...)
}
\arguments{
\item{x}{an object}

\item{...}{dot-dot-dot, ignored}
}
\value{
the object itself
}
\description{
Print a Module
}
