% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{provide}
\alias{provide}
\title{Provide objects from a module}
\usage{
provide(...)
}
\arguments{
\item{...}{dot-dot-dot: name of any object to be accessible by user}
}
\description{
Provide objects from a module
}
\examples{

provide(a)

a <- 1
b <- 2
c <- 3
}
