% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/declaratives.R
\name{provide}
\alias{provide}
\title{Provide Objects from a Module}
\usage{
provide(...)
}
\arguments{
\item{...}{name of any object to be accessible by user; name or character}
}
\description{
Can only be used inside a module expression.
If this function is used, only the names included as argument are public.
If not used, every name in the module will be public.
}
\examples{

\dontrun{
provide(a, c)

a <- 1
b <- 2
c <- 3
d <- 4
}

}
