% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R, R/namespace.R
\name{provide}
\alias{provide}
\title{Provide objects from a module}
\usage{
provide(...)

provide(...)
}
\arguments{
\item{...}{dot-dot-dot}

\item{...}{dot-dot-dot}
}
\description{
Provide objects from a module

Provide objects from a module
}
