% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/declaratives.R
\name{refer}
\alias{refer}
\title{Refer bindings from a module to another}
\usage{
refer(..., include = c(), exclude = c(), prefix = "", sep = ".")
}
\arguments{
\item{...}{names of modules; dot-dot-dot}

\item{include}{names to include; character vector}

\item{exclude}{names to excludde; character vector}

\item{prefix}{prefix to names; character}

\item{sep}{separator between prefix and names; character}
}
\description{
Only legal inside a module context
}
