% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/declaratives.R
\name{refer}
\alias{refer}
\title{Copy Bindings from a Module to Another}
\usage{
refer(..., include = c(), exclude = c(), prefix = "", sep = ".")
}
\arguments{
\item{...}{names of modules; dot-dot-dot}

\item{include}{names to include; character vector}

\item{exclude}{names to excludde; character vector}

\item{prefix}{prefix to names; character}

\item{sep}{separator between prefix and names; character}
}
\description{
Can only be used inside a module expression.
Makes reference to objects from one module, with specified filters.
}
