% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R, R/namespace.R
\name{refer}
\alias{refer}
\title{Exposes the objects in one environment to another}
\usage{
refer(source, target = parent.frame())

refer(source, target = parent.frame())
}
\arguments{
\item{source}{an environment that provides the objects}

\item{target}{an environment that receives the objects}

\item{source}{an environment that provides the objects}

\item{target}{an environment that receives the objects}
}
\description{
Exposes the objects in one environment to another

Exposes the objects in one environment to another
}
