% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/declaratives.R
\name{require}
\alias{require}
\title{Load/Attach Package to Local Search Path}
\usage{
require(package)
}
\arguments{
\item{package}{name of the package; name or character}
}
\description{
Can only be used in a module expression.
Emulates the effect of base::require() in its containing module, making functions and their chain of environment availab.e
 Masks base::require() inside a module context.
}
\examples{

mod_tcl <- mod::ule({
    require(tcltk)
    f <- tcl
})

identical(mod_tcl$f, tcltk::tcl)

}
