% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/declaratives.R
\name{require}
\alias{require}
\title{Attach package to local search path}
\usage{
require(package)
}
\arguments{
\item{package}{name of the package; name or character string}
}
\description{
Attach package to local search path
}
