% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{use}
\alias{use}
\title{Use a module}
\usage{
use(module, as)
}
\arguments{
\item{module}{path to an R file or a symbol for a module object}

\item{as}{name to be used in the search path}
}
\description{
Load and attach a module to the search path
}
\examples{
\dontrun{
use("~/R/example_module")
}
}
